Split markdown function into three pieces:  - parse_references  - parse_notes (takes references as parameter)  - parse_markdown (takes references and notes as parameter)
Added process_raw_blocks function, which goes through the element tree
and calls parse_markdown again, as needed, for RAW blocks.
This is no longer done in markdown_output functions.

This fixes a bug in which reference-style links would not work in
blockquotes and blocks embedded in lists.  It is also a cleaner
design.

This patch also includes some whitespace cleanups.