Scan for setext pattern before parsing Inlines.
Closes Issue #3.

Previously, certain kinds of input (e.g. Emphasis.txt from the
MDTest package) would send peg-markdown into a morass of backtracking
in parsing a Setext header (which begins with a list of Inlines).

We now check to make sure the text has the shape of a setext header
before parsing any Inlines.

This yields large performance gains across the board!
Many thanks to knieriem for pointing out the problem and suggesting a
(somewhat different) fix.