Fixed exponential slowdown for some HTML.
Thanks to sirthias for both the bug report and the fix.

sirthias notes: Rather than recursing into "HtmlBlockInTags" in each sub rule
underneath "HtmlBlockInTags" the parser should recurse only in HTML blocks
that have the same tag name as the enclosing tag. Recursion is only required
to match corresponding pairs of opening/closing tags, not for rebuilding the
HTML tag structure as such.