Add strike-through support for groff output
A groff macro is defined to strike through words enclosed in pairs of
`~~` when rendering to the Postscript device. The macro doesn't yield
suitable results when rendering to any device other than Postscript.

As the macro application requires that the string be placed on its own
line, leading dots have to be escaped that result from splitting strings
like `~~Hello~~. Goodbye` after the last `~~`. Without escaping, the
groff output would be

    .ST "Hello"
    . Goodbye

leading to a failure to process ". Goodbye".